+++ packages/components/src/components.d.ts
@@ -8,7 +8,6 @@ import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { Events, Props } from "./types";
 import { BalMode } from "./utils/config";
 import { BalCarouselItemData } from "./components/bal-carousel/bal-carousel.type";
-import { BannerStatusContext } from "./components/docs/bal-doc-banner-status/bal-doc-banner-status";
 import { FileUploadRejectedFile } from "./components/form/bal-file-upload/bal-file-upload.type";
 import { OverlayEventDetail } from "./components/notice/bal-modal/bal-modal.type";
 import { LevelInfo } from "./components/bal-navigation/utils/level.utils";
@@ -628,9 +627,6 @@ export namespace Components {
     interface BalDocBanner {
         "subtitle": string;
     }
-    interface BalDocBannerStatus {
-        "context": BannerStatusContext[];
-    }
     interface BalDocColor {
         "background": boolean;
         "color": string;
@@ -1004,18 +1000,6 @@ export namespace Components {
           * If `true` the icon is rotated 180deg
          */
         "turn": boolean;
-    }
-    interface BalImageSlider {
-        /**
-          * value to set the images aspect ratio
-         */
-        "aspectRatio"?: '1by1' | '3by2' | '4by3' | '16by9';
-    }
-    interface BalImageSliderItem {
-        /**
-          * Src path to the image
-         */
-        "src"?: string;
     }
     interface BalInput {
         /**
@@ -1870,50 +1854,6 @@ export namespace Components {
          */
         "spaceless": boolean;
     }
-    interface BalProductSlider {
-    }
-    interface BalProductSliderItem {
-        /**
-          * Color of the background
-         */
-        "color"?: Props.BalProductSliderItemColor;
-        /**
-          * This attribute instructs browsers to download a URL instead of navigating to it, so the user will be prompted to save it as a local file. If the attribute has a value, it is used as the pre-filled file name in the Save prompt (the user can still change the file name if they want).
-         */
-        "download"?: string;
-        /**
-          * The type of button.
-         */
-        "elementType": Props.BalButtonElementType;
-        /**
-          * Specifies the URL of the page the link goes to
-         */
-        "href"?: string;
-        /**
-          * Label or title of the product
-         */
-        "label"?: string;
-        /**
-          * The name of the button, which is submitted with the form data.
-         */
-        "name"?: string;
-        /**
-          * Specifies the relationship of the target object to the link object. The value is a space-separated list of [link types](https://developer.mozilla.org/en-US/docs/Web/HTML/Link_types).
-         */
-        "rel"?: string;
-        /**
-          * Src path to the image
-         */
-        "src"?: string;
-        /**
-          * Specifies where to display the linked URL. Only applies when an `href` is provided.
-         */
-        "target": Props.BalButtonTarget;
-        /**
-          * The value of the button, which is submitted with the form data.
-         */
-        "value"?: string | number;
-    }
     interface BalRadio {
         /**
           * If `true`, the element is not mutable, focusable, or even submitted with the form. The user can neither edit nor focus on the control, nor its form control descendants.
@@ -2161,14 +2101,6 @@ export namespace Components {
           * Defines content width of the stage
          */
         "containerSize": Props.BalSheetContainer;
-    }
-    interface BalSlider {
-    }
-    interface BalSliderItem {
-        /**
-          * Label of the slide which will be used for pagination tabs
-         */
-        "label": string;
     }
     interface BalSnackbar {
         /**
@@ -2700,10 +2632,6 @@ export interface BalPopoverCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLBalPopoverElement;
 }
-export interface BalProductSliderItemCustomEvent<T> extends CustomEvent<T> {
-    detail: T;
-    target: HTMLBalProductSliderItemElement;
-}
 export interface BalRadioCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLBalRadioElement;
@@ -2878,12 +2806,6 @@ declare global {
     var HTMLBalDocBannerElement: {
         prototype: HTMLBalDocBannerElement;
         new (): HTMLBalDocBannerElement;
-    };
-    interface HTMLBalDocBannerStatusElement extends Components.BalDocBannerStatus, HTMLStencilElement {
-    }
-    var HTMLBalDocBannerStatusElement: {
-        prototype: HTMLBalDocBannerStatusElement;
-        new (): HTMLBalDocBannerStatusElement;
     };
     interface HTMLBalDocColorElement extends Components.BalDocColor, HTMLStencilElement {
     }
@@ -3106,18 +3028,6 @@ declare global {
     var HTMLBalIconElement: {
         prototype: HTMLBalIconElement;
         new (): HTMLBalIconElement;
-    };
-    interface HTMLBalImageSliderElement extends Components.BalImageSlider, HTMLStencilElement {
-    }
-    var HTMLBalImageSliderElement: {
-        prototype: HTMLBalImageSliderElement;
-        new (): HTMLBalImageSliderElement;
-    };
-    interface HTMLBalImageSliderItemElement extends Components.BalImageSliderItem, HTMLStencilElement {
-    }
-    var HTMLBalImageSliderItemElement: {
-        prototype: HTMLBalImageSliderItemElement;
-        new (): HTMLBalImageSliderItemElement;
     };
     interface HTMLBalInputElement extends Components.BalInput, HTMLStencilElement {
     }
@@ -3377,18 +3287,6 @@ declare global {
         prototype: HTMLBalPopoverContentElement;
         new (): HTMLBalPopoverContentElement;
     };
-    interface HTMLBalProductSliderElement extends Components.BalProductSlider, HTMLStencilElement {
-    }
-    var HTMLBalProductSliderElement: {
-        prototype: HTMLBalProductSliderElement;
-        new (): HTMLBalProductSliderElement;
-    };
-    interface HTMLBalProductSliderItemElement extends Components.BalProductSliderItem, HTMLStencilElement {
-    }
-    var HTMLBalProductSliderItemElement: {
-        prototype: HTMLBalProductSliderItemElement;
-        new (): HTMLBalProductSliderItemElement;
-    };
     interface HTMLBalRadioElement extends Components.BalRadio, HTMLStencilElement {
     }
     var HTMLBalRadioElement: {
@@ -3424,18 +3322,6 @@ declare global {
     var HTMLBalSheetElement: {
         prototype: HTMLBalSheetElement;
         new (): HTMLBalSheetElement;
-    };
-    interface HTMLBalSliderElement extends Components.BalSlider, HTMLStencilElement {
-    }
-    var HTMLBalSliderElement: {
-        prototype: HTMLBalSliderElement;
-        new (): HTMLBalSliderElement;
-    };
-    interface HTMLBalSliderItemElement extends Components.BalSliderItem, HTMLStencilElement {
-    }
-    var HTMLBalSliderItemElement: {
-        prototype: HTMLBalSliderItemElement;
-        new (): HTMLBalSliderItemElement;
     };
     interface HTMLBalSnackbarElement extends Components.BalSnackbar, HTMLStencilElement {
     }
@@ -3557,7 +3443,6 @@ declare global {
         "bal-datepicker": HTMLBalDatepickerElement;
         "bal-doc-app": HTMLBalDocAppElement;
         "bal-doc-banner": HTMLBalDocBannerElement;
-        "bal-doc-banner-status": HTMLBalDocBannerStatusElement;
         "bal-doc-color": HTMLBalDocColorElement;
         "bal-doc-download": HTMLBalDocDownloadElement;
         "bal-doc-github": HTMLBalDocGithubElement;
@@ -3595,8 +3480,6 @@ declare global {
         "bal-hint-text": HTMLBalHintTextElement;
         "bal-hint-title": HTMLBalHintTitleElement;
         "bal-icon": HTMLBalIconElement;
-        "bal-image-slider": HTMLBalImageSliderElement;
-        "bal-image-slider-item": HTMLBalImageSliderItemElement;
         "bal-input": HTMLBalInputElement;
         "bal-input-group": HTMLBalInputGroupElement;
         "bal-input-slider": HTMLBalInputSliderElement;
@@ -3640,16 +3523,12 @@ declare global {
         "bal-pagination": HTMLBalPaginationElement;
         "bal-popover": HTMLBalPopoverElement;
         "bal-popover-content": HTMLBalPopoverContentElement;
-        "bal-product-slider": HTMLBalProductSliderElement;
-        "bal-product-slider-item": HTMLBalProductSliderItemElement;
         "bal-radio": HTMLBalRadioElement;
         "bal-radio-group": HTMLBalRadioGroupElement;
         "bal-select": HTMLBalSelectElement;
         "bal-select-option": HTMLBalSelectOptionElement;
         "bal-shape": HTMLBalShapeElement;
         "bal-sheet": HTMLBalSheetElement;
-        "bal-slider": HTMLBalSliderElement;
-        "bal-slider-item": HTMLBalSliderItemElement;
         "bal-snackbar": HTMLBalSnackbarElement;
         "bal-spinner": HTMLBalSpinnerElement;
         "bal-stage": HTMLBalStageElement;
@@ -4314,9 +4193,6 @@ declare namespace LocalJSX {
     interface BalDocBanner {
         "subtitle"?: string;
     }
-    interface BalDocBannerStatus {
-        "context"?: BannerStatusContext[];
-    }
     interface BalDocColor {
         "background"?: boolean;
         "color"?: string;
@@ -4686,18 +4562,6 @@ declare namespace LocalJSX {
           * If `true` the icon is rotated 180deg
          */
         "turn"?: boolean;
-    }
-    interface BalImageSlider {
-        /**
-          * value to set the images aspect ratio
-         */
-        "aspectRatio"?: '1by1' | '3by2' | '4by3' | '16by9';
-    }
-    interface BalImageSliderItem {
-        /**
-          * Src path to the image
-         */
-        "src"?: string;
     }
     interface BalInput {
         /**
@@ -5600,62 +5464,6 @@ declare namespace LocalJSX {
          */
         "spaceless"?: boolean;
     }
-    interface BalProductSlider {
-    }
-    interface BalProductSliderItem {
-        /**
-          * Color of the background
-         */
-        "color"?: Props.BalProductSliderItemColor;
-        /**
-          * This attribute instructs browsers to download a URL instead of navigating to it, so the user will be prompted to save it as a local file. If the attribute has a value, it is used as the pre-filled file name in the Save prompt (the user can still change the file name if they want).
-         */
-        "download"?: string;
-        /**
-          * The type of button.
-         */
-        "elementType"?: Props.BalButtonElementType;
-        /**
-          * Specifies the URL of the page the link goes to
-         */
-        "href"?: string;
-        /**
-          * Label or title of the product
-         */
-        "label"?: string;
-        /**
-          * The name of the button, which is submitted with the form data.
-         */
-        "name"?: string;
-        /**
-          * Emitted when the button loses focus.
-         */
-        "onBalBlur"?: (event: BalProductSliderItemCustomEvent<void>) => void;
-        /**
-          * Emitted when the button has focus.
-         */
-        "onBalFocus"?: (event: BalProductSliderItemCustomEvent<void>) => void;
-        /**
-          * Emitted when the link element has clicked.
-         */
-        "onBalNavigate"?: (event: BalProductSliderItemCustomEvent<MouseEvent>) => void;
-        /**
-          * Specifies the relationship of the target object to the link object. The value is a space-separated list of [link types](https://developer.mozilla.org/en-US/docs/Web/HTML/Link_types).
-         */
-        "rel"?: string;
-        /**
-          * Src path to the image
-         */
-        "src"?: string;
-        /**
-          * Specifies where to display the linked URL. Only applies when an `href` is provided.
-         */
-        "target"?: Props.BalButtonTarget;
-        /**
-          * The value of the button, which is submitted with the form data.
-         */
-        "value"?: string | number;
-    }
     interface BalRadio {
         /**
           * If `true`, the element is not mutable, focusable, or even submitted with the form. The user can neither edit nor focus on the control, nor its form control descendants.
@@ -5928,14 +5736,6 @@ declare namespace LocalJSX {
           * Defines content width of the stage
          */
         "containerSize"?: Props.BalSheetContainer;
-    }
-    interface BalSlider {
-    }
-    interface BalSliderItem {
-        /**
-          * Label of the slide which will be used for pagination tabs
-         */
-        "label"?: string;
     }
     interface BalSnackbar {
         /**
@@ -6397,7 +6197,6 @@ declare namespace LocalJSX {
         "bal-datepicker": BalDatepicker;
         "bal-doc-app": BalDocApp;
         "bal-doc-banner": BalDocBanner;
-        "bal-doc-banner-status": BalDocBannerStatus;
         "bal-doc-color": BalDocColor;
         "bal-doc-download": BalDocDownload;
         "bal-doc-github": BalDocGithub;
@@ -6435,8 +6234,6 @@ declare namespace LocalJSX {
         "bal-hint-text": BalHintText;
         "bal-hint-title": BalHintTitle;
         "bal-icon": BalIcon;
-        "bal-image-slider": BalImageSlider;
-        "bal-image-slider-item": BalImageSliderItem;
         "bal-input": BalInput;
         "bal-input-group": BalInputGroup;
         "bal-input-slider": BalInputSlider;
@@ -6480,16 +6277,12 @@ declare namespace LocalJSX {
         "bal-pagination": BalPagination;
         "bal-popover": BalPopover;
         "bal-popover-content": BalPopoverContent;
-        "bal-product-slider": BalProductSlider;
-        "bal-product-slider-item": BalProductSliderItem;
         "bal-radio": BalRadio;
         "bal-radio-group": BalRadioGroup;
         "bal-select": BalSelect;
         "bal-select-option": BalSelectOption;
         "bal-shape": BalShape;
         "bal-sheet": BalSheet;
-        "bal-slider": BalSlider;
-        "bal-slider-item": BalSliderItem;
         "bal-snackbar": BalSnackbar;
         "bal-spinner": BalSpinner;
         "bal-stage": BalStage;
@@ -6535,7 +6328,6 @@ declare module "@stencil/core" {
             "bal-datepicker": LocalJSX.BalDatepicker & JSXBase.HTMLAttributes<HTMLBalDatepickerElement>;
             "bal-doc-app": LocalJSX.BalDocApp & JSXBase.HTMLAttributes<HTMLBalDocAppElement>;
             "bal-doc-banner": LocalJSX.BalDocBanner & JSXBase.HTMLAttributes<HTMLBalDocBannerElement>;
-            "bal-doc-banner-status": LocalJSX.BalDocBannerStatus & JSXBase.HTMLAttributes<HTMLBalDocBannerStatusElement>;
             "bal-doc-color": LocalJSX.BalDocColor & JSXBase.HTMLAttributes<HTMLBalDocColorElement>;
             "bal-doc-download": LocalJSX.BalDocDownload & JSXBase.HTMLAttributes<HTMLBalDocDownloadElement>;
             "bal-doc-github": LocalJSX.BalDocGithub & JSXBase.HTMLAttributes<HTMLBalDocGithubElement>;
@@ -6573,8 +6365,6 @@ declare module "@stencil/core" {
             "bal-hint-text": LocalJSX.BalHintText & JSXBase.HTMLAttributes<HTMLBalHintTextElement>;
             "bal-hint-title": LocalJSX.BalHintTitle & JSXBase.HTMLAttributes<HTMLBalHintTitleElement>;
             "bal-icon": LocalJSX.BalIcon & JSXBase.HTMLAttributes<HTMLBalIconElement>;
-            "bal-image-slider": LocalJSX.BalImageSlider & JSXBase.HTMLAttributes<HTMLBalImageSliderElement>;
-            "bal-image-slider-item": LocalJSX.BalImageSliderItem & JSXBase.HTMLAttributes<HTMLBalImageSliderItemElement>;
             "bal-input": LocalJSX.BalInput & JSXBase.HTMLAttributes<HTMLBalInputElement>;
             "bal-input-group": LocalJSX.BalInputGroup & JSXBase.HTMLAttributes<HTMLBalInputGroupElement>;
             "bal-input-slider": LocalJSX.BalInputSlider & JSXBase.HTMLAttributes<HTMLBalInputSliderElement>;
@@ -6618,16 +6408,12 @@ declare module "@stencil/core" {
             "bal-pagination": LocalJSX.BalPagination & JSXBase.HTMLAttributes<HTMLBalPaginationElement>;
             "bal-popover": LocalJSX.BalPopover & JSXBase.HTMLAttributes<HTMLBalPopoverElement>;
             "bal-popover-content": LocalJSX.BalPopoverContent & JSXBase.HTMLAttributes<HTMLBalPopoverContentElement>;
-            "bal-product-slider": LocalJSX.BalProductSlider & JSXBase.HTMLAttributes<HTMLBalProductSliderElement>;
-            "bal-product-slider-item": LocalJSX.BalProductSliderItem & JSXBase.HTMLAttributes<HTMLBalProductSliderItemElement>;
             "bal-radio": LocalJSX.BalRadio & JSXBase.HTMLAttributes<HTMLBalRadioElement>;
             "bal-radio-group": LocalJSX.BalRadioGroup & JSXBase.HTMLAttributes<HTMLBalRadioGroupElement>;
             "bal-select": LocalJSX.BalSelect & JSXBase.HTMLAttributes<HTMLBalSelectElement>;
             "bal-select-option": LocalJSX.BalSelectOption & JSXBase.HTMLAttributes<HTMLBalSelectOptionElement>;
             "bal-shape": LocalJSX.BalShape & JSXBase.HTMLAttributes<HTMLBalShapeElement>;
             "bal-sheet": LocalJSX.BalSheet & JSXBase.HTMLAttributes<HTMLBalSheetElement>;
-            "bal-slider": LocalJSX.BalSlider & JSXBase.HTMLAttributes<HTMLBalSliderElement>;
-            "bal-slider-item": LocalJSX.BalSliderItem & JSXBase.HTMLAttributes<HTMLBalSliderItemElement>;
             "bal-snackbar": LocalJSX.BalSnackbar & JSXBase.HTMLAttributes<HTMLBalSnackbarElement>;
             "bal-spinner": LocalJSX.BalSpinner & JSXBase.HTMLAttributes<HTMLBalSpinnerElement>;
             "bal-stage": LocalJSX.BalStage & JSXBase.HTMLAttributes<HTMLBalStageElement>;